Fixing a misinterpretation of the parameter representation
Because SpectralMixtureKernel parameters are all constrained, we have to output the constrained parameters. However, internally they are represented by parameters that are allowed to vary in the (-inf, +inf) range, called raw_parameter_name, which is the actual parameter, and then the constraint is applied to the raw parameter and fed to the model for the forward pass. Hence, we need to make sure we output the constrained parameters, not the raw ones! However, the constrained parameters don't seem to show up by name in any of the model attributes, which makes it rather irritating to find them.