Improve getting and printing parameters
Previously, getting the values of the hyperparameters was a pain - you could automatically get the values of the raw hyperparameters (i.e. the values that are allowed to vary in the (-inf, inf) space, before they have been transformed to the constrained space), but those are often difficult to interpret. Alternatively, if you knew exactly what to ask for, you could get the values of the constrained parameters (i.e. the ones with meaningful values in the data space), but indeed you had to know ahead of time exactly what they are called. This horrible hack provides a way around that. Next step is to make automatically transform back to the raw data space, as presently they will be in the transformed data space.