Refactoring `fit()` to make it cleaner
Several long segments of boilerplate have been extracted to their own methods. This way the routines are shorter and simpler, and the `set_` methods can be called independently to setup a model or likelihood ahead of time if the user prefers. This also opens up the flexibility to manually set the `self.likelihood` or `model` attributes if the user so desires.