Exploiting `Lightcurve` objects in `train()`
You can now pass a `Lightcurve` object to `train()`. This means that `Lightcuve.fit()` only needs to pass `self` to `train()` instead of several arguments, and means that the hyperparameters can be recorded in the input space rather than the transformed space! I also took this chance to improve a number of PEP8 compliance aspects of `train()`.