Providing means to set parameter constraints
PGMUVI will now accept parameter constraints and adjust them to deal with the transformed space. You can use `Lightcurve.set_constraint()` to set one or more constraints using this. It will also apply some reasonable sensible default constraints - e.g. prevent any of the `mixture_means` from being a frequency that is too low to fit in the time sampled by the data, and forcing the mean to be within the range of the y data.