Implementing postprocessing for MCMC runs
Routines are now implemented to generate summary statistics, trace plots, and corner plots, although these are untested. There is a partial implementation for the posterior predictive plot, which is also untested. Still TBD is plotting MCMC samples of the PSD. Big caveat - MCMC samples are still in the transformed frame, not the input frame, and so the InferenceData object needs to be manipulated a bit before continuing.