(Hopefully) complete MCMC implementation
This is now a complete implementation of using MCMC to interpret lightcurves. The example notebook still needs a bit of polishing, and the descriptive parts need to be filled in, but it is (almost) feature complete.
The one and only missing feature is to be able to run multiple chains - at the moment this will crash with no protection, but in future we should make it iteratively do MCMC, concatenate the results to the InferenceData object and be able to plot results from multiple chains in the plots. This is work for the future, however, I think this is a good enough start.