Starting on contributor guide
This is quite a generic attempt, but includes both a code of conduct and some general guidelines for contributing to pgmuvi. I don't think this does enough to emphasise that you don't need to worry about impostor syndrome, we want to work with people to help them improve. It could also potentially use more improvement on the Code of Conduct?