Updating from setup.py to pyproject.toml
This brings the packaging for pgmuvi up to modern standards, more or less. There is a minimal setup.py to maintain compatability with older versions of pip, but setuptools will find the info from pyproject in that case. This also should ensure that editable installs work smoothly most of the time. This has also made it possible to easily update the version to be dynamically determined at build time from the git tags, commit hash and whether there are any uncommitted changes (ditry). This should make it easier to track versioning.